+++ RELEASENOTES.md
@@ -19,6 +19,7 @@ New features, fixed bugs, known defects and other noteworthy changes to each rel
 * Feature - Show country information in the Partner Network table
 * Feature - Show CX Member info in the Partner Network table
 * Feature - App Release Process Step 6
+* Bug - Partner network detail overlay implemented
 
 
 ### 0.5.3